+++ apps/watch/src/components/HomeVideos/Card/HomeVideoCard.spec.tsx
@@ -1,7 +1,7 @@
 import { ThemeProvider } from '@core/shared/ui/ThemeProvider'
 import { ThemeName, ThemeMode } from '@core/shared/ui/themes'
 import { render } from '@testing-library/react'
-import { videos } from '../../../../pages'
+import { videos } from '../videosData'
 import { data } from '../testData'
 import { HomeVideoCard } from '.'
