+++ apps/watch/src/components/HomeVideos/videosData.ts
@@ -1,0 +1,28 @@
+import { HomeVideo } from './HomeVideos'
+import { FilmType } from './Card'
+
+export const videos: HomeVideo[] = [
+  {
+    id: '1_jf-0-0',
+    designation: FilmType.feature
+  },
+  { id: '2_GOJ-0-0', designation: FilmType.feature },
+  { id: '1_jf6119-0-0', designation: FilmType.feature },
+  { id: '1_wl604423-0-0', designation: FilmType.feature },
+  { id: 'MAG1', designation: FilmType.feature },
+  { id: '1_wl7-0-0', designation: FilmType.series },
+  { id: '3_0-8DWJ-WIJ_06-0-0', designation: FilmType.feature },
+  { id: '2_Acts-0-0', designation: FilmType.feature },
+  { id: '2_GOJ4904-0-0', designation: FilmType.feature },
+  // TODO: LUMO collection goes here
+  { id: '2_Acts7331-0-0', designation: FilmType.feature },
+  { id: '3_0-8DWJ-WIJ', designation: FilmType.feature },
+  { id: '2_ChosenWitness', designation: FilmType.animation },
+  // TODO: LUMO collection gospel of luke here
+  { id: '1_cl1309-0-0', designation: FilmType.feature },
+  { id: '1_jf6102-0-0', designation: FilmType.feature },
+  { id: '2_0-FallingPlates', designation: FilmType.series },
+  { id: '2_Acts7345-0-0', designation: FilmType.feature },
+  { id: '1_mld-0-0', designation: FilmType.feature },
+  { id: '1_jf6101-0-0', designation: FilmType.feature }
+]
